requirements.txt
openai>=1.0.0
tweepy>=4.14.0
python-linkedin-v2>=0.8.0
instagrapi>=1.16.0
schedule>=1.2.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
